complete the part of add new post,set loading , and save in firebase

+++ utils/fetchLocalStorage.js
@@ -1,8 +1,9 @@
 export const fetchUser = () => {
+    if (typeof window !== 'undefined') {
+        const userInfo = localStorage.getItem('user') !== 'undifined' ?
+            JSON.parse(localStorage.getItem('user')) :
+            localStorage.clear();
+            return userInfo;
+    }
 
-    const userInfo = localStorage.getItem('user') !== 'undifined' ?
-        JSON.parse(localStorage.getItem('user')) :
-        localStorage.clear();
-
-    return userInfo;
 };
